moving providers init test that i added to a different file (not __init__.py)
